requirements.txt
Flask==2.1.2
dnspython==2.2.0